pyproject.toml
[project]
name = "explainlike5"
version = "0.1.0"
description = "Explain or comment Python functions using LLMs like you're 5 years old."
authors = [
    {name = "Karim",email = "115158216+karimdiabb@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.95.1,<2.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "explainlike5"}]

[tool.poetry.scripts]
explainlike5 = "explainlike5.cli:main"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
black = "^25.1.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
